solved part 2 with math and huge hint
I realized that the solution had to do something with the last
three bits of A since the general pattern was bitshifting by 3
some number of times, but I couldn't come up with the definitive
pattern. Turns out you can actually go in reverse and keep multiplying
by 3 every time the ends of the instructions and the new program
match, and that reduces the search space quite a bit.

Subreddit came in clutch!